+++ intcode/src/main/kotlin/com/cbob/intcode/commands/AdjustRelativeBaseCommand.kt
@@ -6,7 +6,7 @@ import com.cbob.intcode.State
 class AdjustRelativeBaseCommand : Command {
     override fun execute(state: State, params: CommandParams): Boolean {
         val offset = state.instructions[state.instPtr + 1]
-        state.relBase += offset
+        state.relBase += offset.toInt()
 
         state.instPtr += 2
         return true
