+++ intcode/src/main/kotlin/com/cbob/intcode/commands/Command.kt
@@ -6,16 +6,16 @@ import com.cbob.intcode.State
 interface Command {
     fun execute(state: State, params: CommandParams): Boolean
 
-    fun getVal(state: State, idx: Int, paramMode : CommandParams.ParamMode) : Int {
+    fun getVal(state: State, idx: Int, paramMode : CommandParams.ParamMode) : Long {
         val currInstr = state.instructions[idx]
         return when (paramMode) {
             CommandParams.ParamMode.Position -> {
-                state.instructions[currInstr]
+                state.instructions[currInstr.toInt()]
             }
             CommandParams.ParamMode.Immediate -> currInstr
             CommandParams.ParamMode.Relative -> {
                 val newAddr = state.relBase + currInstr
-                state.instructions[newAddr]
+                state.instructions[newAddr.toInt()]
             }
         }
     }
